bug 618379 r=aki updated cleanup.py, installApp.py and reboot.py to work with new return values from dm reboot.py depends on local patch to talos devicemanager.py to allow for ipAddr/ipPort parameters

+++ sut_tools/cleanup.py
@@ -34,11 +34,11 @@ if devRoot is None or devRoot == '/tests':
     time.sleep(30)
     sys.exit(1)
 
-status = dm.removeDir(devRoot)
-
-print "removeDir() returned [%s]" % status
-#if status is None:
-#    print "Remote Device Error: all to removeDir() failed"
-#    setFlag(errorFile)
-#    time.sleep(30)
-#    sys.exit(1)
+if dm.dirExists(devRoot):
+    status = dm.removeDir(devRoot)
+    print "removeDir() returned [%s]" % status
+    if status is None or not status:
+       print "Remote Device Error: all to removeDir() failed"
+       setFlag(errorFile)
+       time.sleep(30)
+       sys.exit(1)
